chkcc/pyproject.toml
[project]
name = "chkcc"
version = "0.1.0"
description = "Checkpoint compress CLI for coihuin-compress"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
chkcc = "chkcc.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
  "*.py",
]

[tool.hatch.build.targets.wheel.force-include]
"data" = "chkcc/data"

[tool.hatch.build.targets.wheel.sources]
"" = "chkcc"

[tool.hatch.build]
dev-mode-dirs = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
